Submission

+++ vm_functions.c
@@ -131,7 +131,7 @@ void srl(blob* p_vm, uint32_t instruction) {
     uint32_t rs1 = get_number(instruction, 15, 5);
     uint32_t rs2 = get_number(instruction, 20, 5);
     if (rd != 0)
-        p_vm->registers[rd] = p_vm->registers[rs1] >> p_vm->registers[rs2];
+        p_vm->registers[rd] = ((uint32_t) p_vm->registers[rs1]) >> p_vm->registers[rs2];
     p_vm->PC += 4;
     // printf("srl: rd: %d, rs1: %d, rs2: %d\n", p_vm->registers[rd], p_vm->registers[rs1], p_vm->registers[rs2]);
 }
@@ -140,8 +140,8 @@ void sra(blob* p_vm, uint32_t instruction) {
     uint32_t rd = get_number(instruction, 7, 5);
     uint32_t rs1 = get_number(instruction, 15, 5);
     uint32_t rs2 = get_number(instruction, 20, 5);
-    int32_t shifted = ((uint32_t) p_vm->registers[rs1]) >> p_vm->registers[rs2];
-    int32_t rot_bits = ((uint32_t) p_vm->registers[rs1]) << (32 - p_vm->registers[rs2]);
+    int32_t shifted = p_vm->registers[rs1] >> p_vm->registers[rs2];
+    int32_t rot_bits = p_vm->registers[rs1] << (32 - p_vm->registers[rs2]);
     int32_t combined = shifted | rot_bits;
     if (rd != 0)
         p_vm->registers[rd] = combined;
